[FE] REFACTOR: TopNavBar에 있던 웹소켓 구독 페이지로 이동 #245
