+++ src/frontend/src/components/TopNavBar/index.tsx
@@ -1,6 +1,6 @@
 import { useEffect, useState } from 'react';
 import { Link, useNavigate } from 'react-router-dom';
-import axios from 'axios';
+// import axios from 'axios';
 import AddCircleIcon from '@/assets/img/AddCircle.svg';
 import BellIcon from '@/assets/img/Bell.svg';
 import {
@@ -27,12 +27,6 @@ import { friendApi } from '@/api/endpoints/friend/friend.api';
 import { useFriendStore } from '@/stores/useFriendStore';
 import { useNotificationStore } from '@/stores/useNotificationStore';
 
-import { useCurrentRoomStore } from '@/stores/useCurrentRoomStore';
-import { useVideoStore } from '@/stores/useVideoStore';
-
-const API_KEY = import.meta.env.VITE_YOUTUBE_API_KEY as string;
-
-
 export const TopNavBar = () => {
   const navigate = useNavigate();
   const { user, fetchMyProfile, clearProfile } = useUserStore();
@@ -44,10 +38,7 @@ export const TopNavBar = () => {
     resetNotificationCount,
     fetchNotifications,
   } = useNotificationStore();
-  const { connect, subscribeRoomPlaylistUpdate } = useWebSocketStore();
-  const { currentRoom } = useCurrentRoomStore();
-  const { setVideoQueue } = useVideoStore();
-  const roomId = currentRoom?.roomDetails?.roomInfo?.[0]?.roomId;
+  const { connect } = useWebSocketStore();
   const accessToken = useAuthStore(state => state.accessToken);
 
   const [isRoomCreateModalOpen, setIsRoomCreateModalOpen] = useState(false);
@@ -79,70 +70,6 @@ export const TopNavBar = () => {
     };
     initializeUser();
   }, [accessToken, clearProfile, fetchMyProfile, fetchMyRooms, connect]);
-
-  // YouTube API를 이용하여 영상 정보(제목 & 유튜버) 가져오기
-  const fetchVideoDetails = async (videoIds: string[]) => {
-    if (!videoIds.length) return {};
-
-    try {
-      const { data } = await axios.get('https://www.googleapis.com/youtube/v3/videos', {
-        params: {
-          part: 'snippet',
-          id: videoIds.join(','),
-          key: API_KEY,
-          hl: 'ko',
-        },
-      });
-
-      const videoDetailsMap: Record<string, { title: string; youtuber: string }> = {};
-      data.items.forEach(
-        (item: { id: string; snippet: { title: string; channelTitle: string } }) => {
-          videoDetailsMap[item.id] = {
-            title: item.snippet.title,
-            youtuber: item.snippet.channelTitle,
-          };
-        },
-      );
-
-      return videoDetailsMap;
-    } catch (error) {
-      console.error('YouTube API 호출 실패:', error);
-      return {};
-    }
-  };
-
-  // 방의 플레이리스트 업데이트 구독
-  useEffect(() => {
-    if (!roomId) return;
-
-    subscribeRoomPlaylistUpdate(roomId, async data => {
-      console.log('플레이리스트 업데이트 수신:', data);
-
-      if (data?.playlist && Array.isArray(data.playlist)) {
-        const sortedPlaylist = data.playlist.sort((a, b) => a.order - b.order);
-
-        const videoIds = sortedPlaylist
-          .map(item => {
-            return item.url.split('v=')[1]?.split('&')[0] || '';
-          })
-          .filter(id => id);
-
-        const videoDetailsMap = await fetchVideoDetails(videoIds);
-
-        const updatedQueue = sortedPlaylist.map(item => {
-          const videoId = item.url.split('v=')[1]?.split('&')[0] || '';
-          return {
-            id: videoId,
-            start: parseInt(item.url.split('t=')[1] || '0', 10),
-            thumbnail: `https://img.youtube.com/vi/${videoId}/0.jpg`,
-            title: item.title || videoDetailsMap[videoId]?.title || '제목 없음',
-            youtuber: item.youtuber || videoDetailsMap[videoId]?.youtuber || '유튜버 정보 없음',
-          };
-        });
-        setVideoQueue(updatedQueue);
-      }
-    });
-  }, [roomId, subscribeRoomPlaylistUpdate, setVideoQueue]);
 
   const clickCreateRoom = (e: React.MouseEvent<HTMLDivElement>) => {
     e.stopPropagation();
